+++ app/qcl/views/routes.py
@@ -1,21 +1,17 @@
 from qcl import app
-from qcl.utils import code_format, fileops, validate
+from qcl.utils import code_format, fileops
 from qcl.integrations import gpt, linter, testrunner
 from qcl.models.user import User
 from qcl.models import function
+from qcl.views.forms import SignupForm, LoginForm, EmailVerificationForm, CodeForm, DocForm, TestForm, ClassifyForm
 
 from functools import wraps
 import traceback
 import json
-
 from flask import render_template, request, session, redirect, url_for
-from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField, TextAreaField
-from wtforms.validators import DataRequired, Length
 
 import logging
 logging.basicConfig(level=logging.INFO)
-
 
 
 def needs_user(func):
@@ -38,7 +34,6 @@ def needs_admin(func):
         return func(*args, **kwargs)
     return wrapper
 
-
 @app.route("/unauthorized")
 def unauthorized():
     return render_template("unauthorized.html"), 401
@@ -49,76 +44,6 @@ def badrequest():
 @app.route("/forbidden")
 def forbidden():
     return render_template("forbidden.html"), 403
-
-class SignupForm(FlaskForm):
-    username = StringField('Username', validators=[
-        DataRequired("Username cannot be empty"),
-        Length(min=14, max=100),
-        validate.check_email
-
-    ])
-    password = PasswordField('Password', validators=[
-        DataRequired("Password cannot be empty"),
-        Length(min=8, max=100)
-    ])
-
-class LoginForm(FlaskForm):
-    username = StringField('Username', validators=[
-        DataRequired("Username cannot be empty")
-
-    ])
-    password = PasswordField('Password', validators=[
-        DataRequired("Password cannot be empty")
-    ])
-
-class EmailVerificationForm(FlaskForm):
-    code = StringField("Verification code", validators=[
-        DataRequired("The code is required"),
-        Length(min=6, max=12)
-    ])
-
-class CodeForm(FlaskForm):
-    code = StringField(validators=[
-        Length(min=10, max=1500)
-    ])
-    lint = SubmitField("Run Linting")
-    doc = SubmitField("Next")
-
-class DocForm(FlaskForm):
-    code = StringField(validators=[
-        Length(min=10, max=1500)
-    ])
-    documented = StringField(validators=[
-        Length(min=10, max=2000)
-    ])
-    generate = SubmitField("Generate Docs", render_kw={"formnovalidate": True})
-    next = SubmitField("Next")
-
-class TestForm(FlaskForm):
-    documented = StringField(validators=[
-        Length(min=10, max=2000)
-    ])
-    unittests = StringField(validators=[
-        Length(min=10, max=3000)
-    ])
-    generate = SubmitField("Generate Tests", render_kw={"formnovalidate": True})
-    run = SubmitField("Run Tests", render_kw={"formnovalidate": True})
-    next = SubmitField("Next")
-
-class ClassifyForm(FlaskForm):
-    keywords = TextAreaField("Keywords (comma separated)", validators=[
-        Length(min=3, max=200)
-    ])
-    name = StringField("Function name", validators=[
-        DataRequired(),
-        Length(min=3, max=50)
-    ])
-    usecase = TextAreaField("Use case", validators=[
-        DataRequired(),
-        Length(min=10, max=500)
-    ])
-    generate = SubmitField("Generate keywords", render_kw={"formnovalidate": True})
-    save = SubmitField("Save function")
 
 @app.route("/", methods=['GET', 'POST'])
 def index():
